qdrant_embeddings/query_client/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
qdrant-client
